% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdr_calc_geo_dists.R
\name{bdr_calc_geo_dist}
\alias{bdr_calc_geo_dist}
\title{Calculate geographic distance estimates}
\usage{
bdr_calc_geo_dist(
  pnt_lat = NULL,
  pnt_lon = NULL,
  vec_lat = NULL,
  vec_lon = NULL
)
}
\arguments{
\item{pnt_lat}{a point latitude}

\item{pnt_lon}{a point longitude}

\item{vec_lat}{a vector of latitude points}

\item{vec_lon}{a vector of longitude points}
}
\value{
a vector of distances from the point coordinate the length of vec_lat
}
\description{
Calculates geographic distances between a latitude and longitude point location
against vectors of similar latitude and longitude points representing multiple locations
Note: this travel distance estimate is only useful for relatively close distances, i.e. a couple hundred miles
}
\examples{
bdr_calc_geo_dist(pnt_lat = 44.98425, pnt_lon = -93.26644, vec_lat = c(44.96280,44.97361,44.97566), vec_lon = c(-93.34795,-93.22987,-93.27227))
}
