% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdr_get_fdic_branch_data.R
\name{bdr_get_fdic_branch_data}
\alias{bdr_get_fdic_branch_data}
\title{Download and return FDIC Branch Deposit Data for all US branches for any available period}
\usage{
bdr_get_fdic_branch_data(
  .year = 2020,
  save_dir = here::here("data"),
  keep_fdic_rds_file = T,
  keep_fdic_zip_file = F
)
}
\arguments{
\item{.year}{integer, the year you want to get fdic branch data for}

\item{save_dir}{string, the location to save the zip and rds files, default is here::here('data')}

\item{keep_fdic_rds_file}{boolean to save the rds file in the save_dir, default is T}

\item{keep_fdic_zip_file}{boolean to save the zip file in the save_dir, default is F}
}
\value{
tibble of the entire fdic branch data file and corresponding saved rds file, as well as saved zip files if elected
}
\description{
Gets the complete yearly data from the fdic website and saves the data to a
local specified folder, or if the data has already been downloaded, the
function just picks up the saved data
Note: data is only available from 1994 or sooner
human URL: https://www7.fdic.gov/sod/dynaDownload.asp?barItem=6
human URL info: https://www7.fdic.gov/sod/sodDownload3.asp?sState=all&sInfoAsOf=2020&submit1=Continue&barItem=6
}
\examples{
fdic_2020 <- bdr_get_fdic_branch_data(2020)
}
