% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdr_subset_by_dists.R
\name{bdr_subset_by_dists}
\alias{bdr_subset_by_dists}
\title{Subset vector of distances}
\usage{
bdr_subset_by_dists(eucl_dists = NULL, nearest_n = Inf, within_dist = Inf)
}
\arguments{
\item{eucl_dists}{a vector of numeric distances from function calc_eucl_dist()}

\item{nearest_n}{a number indicating how many of the nearest distances to return, default is Inf}

\item{within_dist}{a number specifying the maximum distance in miles used to return observations, default is Inf}
}
\value{
a vector of TRUE or FALSE values denoting which observations are within the nearest_n and within_dist specs
}
\description{
Identifies the subset of distances within x miles of 'within_dist' accounting
for the maximum number of observations to return from "nearest_n".
Function is useful for filter a dataframe by a vector of input distances
Note: nearest_n and within_dist act as 'and' statements, returning the intersection of the pair if both are set
}
\examples{
bdr_subset_by_dists(eucl_dists = c(1, 10.3, 0.5109930, 0.5023643, 0), nearest_n = 4, within_dist = .9 )
}
